Add Next button to media player
and moved from a "build most dom elements in js" pattern to "put as much
as you can in the html file and dynamically do what you need with the
rest" pattern, because it seemed like a lot of unnecessary work to tear
everything down & rebuild so often (e.g. on each song change).